+++ cmake/nfxGraphicsGlSources.cmake
@@ -30,6 +30,9 @@ set(NFX_GRAPHICS_GL_SOURCES
     src/gl/core/functions/Functions_4_6.cpp
     src/gl/core/Context.cpp
 
+    # Shaders
+    src/gl/core/shaders/ShaderProgram.cpp
+
     # Textures
     src/gl/core/textures/Texture2D.cpp
 )